src/interaction/psychological_profile_engine.py
import numpy as np
from typing import Dict, List, Any, Optional, Tuple
from dataclasses import dataclass, field
from sklearn.preprocessing import StandardScaler
from sklearn.decomposition import PCA
from sklearn.cluster import KMeans

@dataclass
class CognitiveStyle:
    """
    Represents a user's cognitive processing style
    """
    analytical_score: float
    intuitive_score: float
    visual_score: float
    verbal_score: float
    sequential_score: float
    global_score: float
    active_score: float
    reflective_score: float

@dataclass
class PersonalityProfile:
    """
    Represents a user's personality traits
    """
    openness: float
    conscientiousness: float
    extraversion: float
    agreeableness: float
    neuroticism: float
    traits: Dict[str, float] = field(default_factory=dict)

@dataclass
class LearningPreference:
    """
    Represents a user's learning preferences
    """
    visual_preference: float
    auditory_preference: float
    kinesthetic_preference: float
    reading_preference: float
    multimodal_score: float

@dataclass
class UserProfile:
    """
    Comprehensive user psychological profile
    """
    user_id: str
    cognitive_style: CognitiveStyle
    personality_profile: PersonalityProfile
    learning_preference: LearningPreference
    interaction_history: List[Dict[str, Any]] = field(default_factory=list)
    adaptation_metrics: Dict[str, float] = field(default_factory=dict)

class PsychologicalProfileEngine:
    def __init__(self):
        """
        Initialize Psychological Profile Engine
        """
        # Profile analysis configuration
        self.profile_config = {
            'min_interaction_history': 10,
            'adaptation_threshold': 0.75,
            'learning_rate': 0.1
        }
        
        # Initialize analysis models
        self.feature_scaler = StandardScaler()
        self.profile_pca = PCA(n_components=5)
        self.style_clustering = KMeans(n_clusters=4)
        
        # Interaction style templates
        self.interaction_templates = {
            'analytical': {
                'detail_level': 'high',
                'explanation_style': 'logical',
                'visualization_preference': 'diagrams'
            },
            'intuitive': {
                'detail_level': 'medium',
                'explanation_style': 'metaphorical',
                'visualization_preference': 'conceptual'
            },
            'visual': {
                'detail_level': 'medium',
                'explanation_style': 'visual',
                'visualization_preference': 'interactive'
            },
            'verbal': {
                'detail_level': 'high',
                'explanation_style': 'narrative',
                'visualization_preference': 'text'
            }
        }
    
    def create_user_profile(
        self,
        user_id: str,
        initial_assessment: Optional[Dict[str, Any]] = None
    ) -> UserProfile:
        """
        Create a new user psychological profile
        
        Args:
            user_id: User identifier
            initial_assessment: Initial psychological assessment data
        
        Returns:
            Created user profile
        """
        # Process initial assessment or use defaults
        assessment = initial_assessment or self._generate_default_assessment()
        
        # Create cognitive style profile
        cognitive_style = CognitiveStyle(
            analytical_score=assessment.get('analytical', 0.5),
            intuitive_score=assessment.get('intuitive', 0.5),
            visual_score=assessment.get('visual', 0.5),
            verbal_score=assessment.get('verbal', 0.5),
            sequential_score=assessment.get('sequential', 0.5),
            global_score=assessment.get('global', 0.5),
            active_score=assessment.get('active', 0.5),
            reflective_score=assessment.get('reflective', 0.5)
        )
        
        # Create personality profile
        personality_profile = PersonalityProfile(
            openness=assessment.get('openness', 0.5),
            conscientiousness=assessment.get('conscientiousness', 0.5),
            extraversion=assessment.get('extraversion', 0.5),
            agreeableness=assessment.get('agreeableness', 0.5),
            neuroticism=assessment.get('neuroticism', 0.5)
        )
        
        # Create learning preference profile
        learning_preference = LearningPreference(
            visual_preference=assessment.get('visual_learning', 0.5),
            auditory_preference=assessment.get('auditory_learning', 0.5),
            kinesthetic_preference=assessment.get('kinesthetic_learning', 0.5),
            reading_preference=assessment.get('reading_learning', 0.5),
            multimodal_score=assessment.get('multimodal_learning', 0.5)
        )
        
        return UserProfile(
            user_id=user_id,
            cognitive_style=cognitive_style,
            personality_profile=personality_profile,
            learning_preference=learning_preference
        )
    
    def update_profile(
        self,
        profile: UserProfile,
        interaction_data: Dict[str, Any]
    ) -> UserProfile:
        """
        Update user profile based on interaction data
        
        Args:
            profile: User profile to update
            interaction_data: New interaction data
        
        Returns:
            Updated user profile
        """
        # Add interaction to history
        profile.interaction_history.append(interaction_data)
        
        # Update cognitive style
        self._update_cognitive_style(profile, interaction_data)
        
        # Update learning preferences
        self._update_learning_preferences(profile, interaction_data)
        
        # Update adaptation metrics
        profile.adaptation_metrics = self._compute_adaptation_metrics(profile)
        
        return profile
    
    def generate_interaction_strategy(
        self,
        profile: UserProfile,
        context: Dict[str, Any]
    ) -> Dict[str, Any]:
        """
        Generate personalized interaction strategy
        
        Args:
            profile: User profile
            context: Interaction context
        
        Returns:
            Personalized interaction strategy
        """
        # Analyze cognitive style and context
        dominant_style = self._determine_dominant_style(profile.cognitive_style)
        
        # Select base interaction template
        strategy = self.interaction_templates[dominant_style].copy()
        
        # Personalize based on learning preferences
        strategy.update(self._personalize_learning_approach(
            profile.learning_preference,
            context
        ))
        
        # Adjust for personality traits
        strategy.update(self._adjust_for_personality(
            profile.personality_profile,
            context
        ))
        
        return strategy
    
    def analyze_interaction_patterns(
        self,
        profile: UserProfile
    ) -> Dict[str, Any]:
        """
        Analyze user interaction patterns
        
        Args:
            profile: User profile
        
        Returns:
            Interaction pattern analysis
        """
        if len(profile.interaction_history) < self.profile_config['min_interaction_history']:
            return {'status': 'insufficient_data'}
        
        # Extract interaction features
        features = self._extract_interaction_features(profile.interaction_history)
        
        # Normalize features
        normalized_features = self.feature_scaler.fit_transform(features)
        
        # Perform PCA
        principal_components = self.profile_pca.fit_transform(normalized_features)
        
        # Cluster interaction styles
        clusters = self.style_clustering.fit_predict(principal_components)
        
        return {
            'dominant_patterns': self._identify_dominant_patterns(clusters),
            'style_evolution': self._analyze_style_evolution(principal_components),
            'adaptation_recommendations': self._generate_adaptation_recommendations(profile)
        }
    
    def _generate_default_assessment(self) -> Dict[str, Any]:
        """
        Generate default psychological assessment
        
        Returns:
            Default assessment values
        """
        return {
            'analytical': 0.5,
            'intuitive': 0.5,
            'visual': 0.5,
            'verbal': 0.5,
            'sequential': 0.5,
            'global': 0.5,
            'active': 0.5,
            'reflective': 0.5,
            'openness': 0.5,
            'conscientiousness': 0.5,
            'extraversion': 0.5,
            'agreeableness': 0.5,
            'neuroticism': 0.5,
            'visual_learning': 0.5,
            'auditory_learning': 0.5,
            'kinesthetic_learning': 0.5,
            'reading_learning': 0.5,
            'multimodal_learning': 0.5
        }
    
    def _update_cognitive_style(
        self,
        profile: UserProfile,
        interaction_data: Dict[str, Any]
    ):
        """
        Update cognitive style based on interaction
        
        Args:
            profile: User profile
            interaction_data: New interaction data
        """
        # Extract cognitive indicators from interaction
        indicators = self._extract_cognitive_indicators(interaction_data)
        
        # Update scores with learning rate
        lr = self.profile_config['learning_rate']
        
        profile.cognitive_style.analytical_score += lr * (indicators.get('analytical', 0) - profile.cognitive_style.analytical_score)
        profile.cognitive_style.intuitive_score += lr * (indicators.get('intuitive', 0) - profile.cognitive_style.intuitive_score)
        profile.cognitive_style.visual_score += lr * (indicators.get('visual', 0) - profile.cognitive_style.visual_score)
        profile.cognitive_style.verbal_score += lr * (indicators.get('verbal', 0) - profile.cognitive_style.verbal_score)
        profile.cognitive_style.sequential_score += lr * (indicators.get('sequential', 0) - profile.cognitive_style.sequential_score)
        profile.cognitive_style.global_score += lr * (indicators.get('global', 0) - profile.cognitive_style.global_score)
        profile.cognitive_style.active_score += lr * (indicators.get('active', 0) - profile.cognitive_style.active_score)
        profile.cognitive_style.reflective_score += lr * (indicators.get('reflective', 0) - profile.cognitive_style.reflective_score)
    
    def _update_learning_preferences(
        self,
        profile: UserProfile,
        interaction_data: Dict[str, Any]
    ):
        """
        Update learning preferences based on interaction
        
        Args:
            profile: User profile
            interaction_data: New interaction data
        """
        # Extract learning indicators
        indicators = self._extract_learning_indicators(interaction_data)
        
        # Update preferences with learning rate
        lr = self.profile_config['learning_rate']
        
        profile.learning_preference.visual_preference += lr * (indicators.get('visual', 0) - profile.learning_preference.visual_preference)
        profile.learning_preference.auditory_preference += lr * (indicators.get('auditory', 0) - profile.learning_preference.auditory_preference)
        profile.learning_preference.kinesthetic_preference += lr * (indicators.get('kinesthetic', 0) - profile.learning_preference.kinesthetic_preference)
        profile.learning_preference.reading_preference += lr * (indicators.get('reading', 0) - profile.learning_preference.reading_preference)
        
        # Update multimodal score
        profile.learning_preference.multimodal_score = self._compute_multimodal_score(profile.learning_preference)
    
    def _determine_dominant_style(self, cognitive_style: CognitiveStyle) -> str:
        """
        Determine dominant cognitive style
        
        Args:
            cognitive_style: User's cognitive style
        
        Returns:
            Dominant style identifier
        """
        style_scores = {
            'analytical': cognitive_style.analytical_score,
            'intuitive': cognitive_style.intuitive_score,
            'visual': cognitive_style.visual_score,
            'verbal': cognitive_style.verbal_score
        }
        
        return max(style_scores.items(), key=lambda x: x[1])[0]
    
    def _personalize_learning_approach(
        self,
        learning_preference: LearningPreference,
        context: Dict[str, Any]
    ) -> Dict[str, Any]:
        """
        Personalize learning approach based on preferences
        
        Args:
            learning_preference: User's learning preferences
            context: Interaction context
        
        Returns:
            Personalized learning approach
        """
        return {
            'content_type': self._select_content_type(learning_preference),
            'interaction_mode': self._select_interaction_mode(learning_preference),
            'presentation_style': self._select_presentation_style(learning_preference, context)
        }
    
    def _adjust_for_personality(
        self,
        personality: PersonalityProfile,
        context: Dict[str, Any]
    ) -> Dict[str, Any]:
        """
        Adjust interaction strategy for personality traits
        
        Args:
            personality: User's personality profile
            context: Interaction context
        
        Returns:
            Personality-adjusted interaction parameters
        """
        return {
            'communication_style': self._select_communication_style(personality),
            'feedback_frequency': self._determine_feedback_frequency(personality),
            'challenge_level': self._determine_challenge_level(personality, context)
        }
    
    def _extract_cognitive_indicators(
        self,
        interaction_data: Dict[str, Any]
    ) -> Dict[str, float]:
        """
        Extract cognitive style indicators from interaction
        
        Args:
            interaction_data: Interaction data
        
        Returns:
            Cognitive style indicators
        """
        # Analyze interaction patterns
        return {
            'analytical': self._compute_analytical_indicator(interaction_data),
            'intuitive': self._compute_intuitive_indicator(interaction_data),
            'visual': self._compute_visual_indicator(interaction_data),
            'verbal': self._compute_verbal_indicator(interaction_data),
            'sequential': self._compute_sequential_indicator(interaction_data),
            'global': self._compute_global_indicator(interaction_data),
            'active': self._compute_active_indicator(interaction_data),
            'reflective': self._compute_reflective_indicator(interaction_data)
        }
    
    def _extract_learning_indicators(
        self,
        interaction_data: Dict[str, Any]
    ) -> Dict[str, float]:
        """
        Extract learning preference indicators from interaction
        
        Args:
            interaction_data: Interaction data
        
        Returns:
            Learning preference indicators
        """
        return {
            'visual': self._compute_visual_learning_indicator(interaction_data),
            'auditory': self._compute_auditory_learning_indicator(interaction_data),
            'kinesthetic': self._compute_kinesthetic_learning_indicator(interaction_data),
            'reading': self._compute_reading_learning_indicator(interaction_data)
        }
    
    def _compute_adaptation_metrics(self, profile: UserProfile) -> Dict[str, float]:
        """
        Compute profile adaptation metrics
        
        Args:
            profile: User profile
        
        Returns:
            Adaptation metrics
        """
        return {
            'style_stability': self._compute_style_stability(profile),
            'learning_efficiency': self._compute_learning_efficiency(profile),
            'interaction_satisfaction': self._compute_satisfaction_score(profile)
        }
    
    # Placeholder implementations for indicator computation methods
    def _compute_analytical_indicator(self, data: Dict[str, Any]) -> float:
        return 0.5  # Placeholder
    
    def _compute_intuitive_indicator(self, data: Dict[str, Any]) -> float:
        return 0.5  # Placeholder
    
    def _compute_visual_indicator(self, data: Dict[str, Any]) -> float:
        return 0.5  # Placeholder
    
    def _compute_verbal_indicator(self, data: Dict[str, Any]) -> float:
        return 0.5  # Placeholder
    
    def _compute_sequential_indicator(self, data: Dict[str, Any]) -> float:
        return 0.5  # Placeholder
    
    def _compute_global_indicator(self, data: Dict[str, Any]) -> float:
        return 0.5  # Placeholder
    
    def _compute_active_indicator(self, data: Dict[str, Any]) -> float:
        return 0.5  # Placeholder
    
    def _compute_reflective_indicator(self, data: Dict[str, Any]) -> float:
        return 0.5  # Placeholder
    
    def _compute_visual_learning_indicator(self, data: Dict[str, Any]) -> float:
        return 0.5  # Placeholder
    
    def _compute_auditory_learning_indicator(self, data: Dict[str, Any]) -> float:
        return 0.5  # Placeholder
    
    def _compute_kinesthetic_learning_indicator(self, data: Dict[str, Any]) -> float:
        return 0.5  # Placeholder
    
    def _compute_reading_learning_indicator(self, data: Dict[str, Any]) -> float:
        return 0.5  # Placeholder
